+++ demo-snippets/vue/install.ts
@@ -1,5 +1,6 @@
 import Vue from 'nativescript-vue';
 import WebView from '@nativescript-community/ui-webview/vue';
+import {WebViewTraceCategory} from '@nativescript-community/ui-webview';
 import {Trace} from '@nativescript/core';
 
 import BasicExample from './BasicExample.vue';