+++ packages/webview-rtc/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.4](https://github.com/nativescript-community/ui-collectionview/compare/v1.4.3...v1.4.4) (2023-02-15)
+
+**Note:** Version bump only for package @nativescript-community/ui-webview-rtc
+
+
+
+
+
+## [1.4.3](https://github.com/nativescript-community/ui-collectionview/compare/v1.4.2...v1.4.3) (2023-01-24)
+
+**Note:** Version bump only for package @nativescript-community/ui-webview-rtc
+
+
+
+
+
 ## [1.4.2](https://github.com/nativescript-community/ui-collectionview/compare/v1.4.1...v1.4.2) (2022-09-02)
 
 **Note:** Version bump only for package @nativescript-community/ui-webview-rtc